Biomes

Cubiomes.Biomes — Module

Biomes

Minecraft biome constants and functions to work with them / compare them.

The list of all biome names (Biomes.biome_name):

ocean

plains

desert

mountains

forest

taiga

swamp

river

nether_wastes

the_end

frozen_ocean

frozen_river

snowy_tundra

snowy_mountains

mushroom_fields

mushroom_field_shore

beach

desert_hills

wooded_hills

taiga_hills

mountain_edge

jungle

jungle_hills

jungle_edge

deep_ocean

stone_shore

snowy_beach

birch_forest

birch_forest_hills

dark_forest

snowy_taiga

snowy_taiga_hills

giant_tree_taiga

giant_tree_taiga_hills

wooded_mountains

savanna

savanna_plateau

badlands

wooded_badlands_plateau

badlands_plateau

small_end_islands

end_midlands

end_highlands

end_barrens

warm_ocean

lukewarm_ocean

cold_ocean

deep_warm_ocean

deep_lukewarm_ocean

deep_cold_ocean

deep_frozen_ocean

seasonal_forest

rainforest

shrubland

the_void

sunflower_plains

desert_lakes

gravelly_mountains

flower_forest

taiga_mountains

swamp_hills

ice_spikes

modified_jungle

modified_jungle_edge

tall_birch_forest

tall_birch_hills

dark_forest_hills

snowy_taiga_mountains

giant_spruce_taiga

giant_spruce_taiga_hills

modified_gravelly_mountains

shattered_savanna

shattered_savanna_plateau

eroded_badlands

modified_wooded_badlands_plateau

modified_badlands_plateau

bamboo_jungle

bamboo_jungle_hills

soul_sand_valley

crimson_forest

warped_forest

basalt_deltas

dripstone_caves

lush_caves

meadow

grove

snowy_slopes

jagged_peaks

frozen_peaks

stony_peaks

old_growth_birch_forest

old_growth_pine_taiga

old_growth_spruce_taiga

snowy_plains

sparse_jungle

stony_shore

windswept_hills

windswept_forest

windswept_gravelly_hills

windswept_savanna

wooded_badlands

deep_dark

mangrove_swamp

cherry_grove

And a special BIOME_NONE.

Index

Cubiomes.Biomes.are_similar

Cubiomes.Biomes.biome_exists

Cubiomes.Biomes.category

Cubiomes.Biomes.is_overworld

Cubiomes.Biomes.mutated

API

Cubiomes.Biomes.are_similar — Method

are_similar(version::MCVersion, biome1::Biome, biome2::Biome)

For a given version, check if two biomes have the same category. wooded_badlands_plateau and badlands_plateau are considered similar even though they have a different category in version <= 1.15.

Cubiomes.Biomes.biome_exists — Method

biome_exists(biome::Biome, version::MCVersion)

Return true if the given biome exists in the given version and false otherwise.

Cubiomes.Biomes.category — Method

category(biome::Biome, version::MCVersion)

Return the category of the given biome in the given version. The categories are:

beach

desert

mountains

forest

snowy_tundra

jungle

mesa

mushroom_fields

stone_shore

ocean

plains

river

savanna

swamp

taiga

nether_wastes

If the biome does not belong to any of these categories, return BIOME_NONE.

Cubiomes.Biomes.is_overworld — Method

is_overworld(biome::Biome, version::MCVersion)

Return true if the given biome is an overworld biome and false otherwise. If the biome does not exist in the given version, return false.

Examples

julia> is_overworld(Biomes.ocean, mcv"1.16",)
true

julia> [biome for biome in instances(Biome) if is_overworld(biome, mcv"1.16")]
90-element Vector{Biome}:
 ocean::Biome = 0x00
 plains::Biome = 0x01
 desert::Biome = 0x02
 ...

Cubiomes.Biomes.mutated — Method

mutated(biome::Biome, version::MCVersion)

Return the mutated variant of the given biome in the given version. If the biome does not have a mutated variant, return BIOME_NONE.

Private API