SeedUtils

Cubiomes.SeedUtils — Module

Minecraft Seed Utilities, like the LCG algorithm used in the seed generation.

Index

Cubiomes.SeedUtils.MAGIC_LCG_INCREMENTOR

Cubiomes.SeedUtils.MAGIC_LCG_MULTIPLIER

Cubiomes.SeedUtils.mc_step_seed

API

Private API

Cubiomes.SeedUtils.MAGIC_LCG_INCREMENTOR — Constant

MAGIC_LCG_INCREMENTOR::UInt64

The incrementor used in the LCG algorithm. This is a constant used in the Minecraft seed generation algorithm.

See Also: MAGIC_LCG_MULTIPLIER, mc_step_seed, LCG wiki

Cubiomes.SeedUtils.MAGIC_LCG_MULTIPLIER — Constant

MAGIC_LCG_MULTIPLIER::UInt64

The multiplier used in the LCG algorithm. This is a constant used in the Minecraft seed generation algorithm.

See Also: MAGIC_LCG_INCREMENTOR, mc_step_seed, LCG wiki

Cubiomes.SeedUtils.mc_step_seed — Method

mc_step_seed(seed::UInt64, salt::UInt64)

Used to generate the next seed in the Minecraft seed generation algorithm, given the current seed and a salt.